requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.13.0
bcrypt==3.2.0
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.15.0
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
numpy==1.21.4
opencv-python==4.5.4.60
pandas==1.3.5
pdf2image==1.16.0
pefile==2021.9.3
Pillow==8.4.0
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
six==1.16.0
tinyaes==1.0.2
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
